COMP: removed size_t to int convesrion warning

+++ VTK/Graphics/vtkYoungsMaterialInterface.cxx
@@ -343,7 +343,7 @@ int vtkYoungsMaterialInterface::RequestData(vtkInformation *vtkNotUsed(request),
 
   // variables visible by debugger
 
-   int nmat = this->Internals->Materials.size();
+   int nmat = (int) this->Internals->Materials.size();
    int nCellData = input->GetCellData()->GetNumberOfArrays();
    int nPointData = input->GetPointData()->GetNumberOfArrays();
    vtkIdType nCells = input->GetNumberOfCells();
